cmty: add cmty_graph and subgraph
cmty_graph: return a graph of communities (old communities are new
nodes, edges are connections between communities).

subgraph: return a subgraph centered at a specific community.  Useful
for visualization and community-centric analysis.

Both of these methods rely on the use of .nodecmtys(), which should
someday be cached so that it isn't regenerated between different
method calls.